+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/aware/SandboxAware.kt
@@ -7,6 +7,8 @@ import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Internal
 import org.jetbrains.intellij.platform.gradle.Constants
 import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformExtension
+import org.jetbrains.intellij.platform.gradle.utils.asPath
+import java.nio.file.Path
 
 /**
  * The interface provides quick access to the sandbox container and specific directories located within it.
@@ -64,3 +66,9 @@ interface SandboxAware : IntelliJPlatformVersionAware {
     @get:Internal
     val sandboxLogDirectory: DirectoryProperty
 }
+
+/**
+ * Path to a properties file which will be used to configure the frontend process if the IDE is started in Split Mode. 
+ */
+val SandboxAware.frontendPropertiesFilePath: Path
+    get() = sandboxContainerDirectory.asPath.resolve("frontend.properties")