+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/tasks.kt
@@ -21,6 +21,7 @@ import org.jetbrains.intellij.platform.gradle.Constants.Sandbox
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
 import org.jetbrains.intellij.platform.gradle.argumentProviders.IntelliJPlatformArgumentProvider
 import org.jetbrains.intellij.platform.gradle.argumentProviders.SandboxArgumentProvider
+import org.jetbrains.intellij.platform.gradle.argumentProviders.SandboxArgumentProviderSplitModeAware
 import org.jetbrains.intellij.platform.gradle.argumentProviders.SplitModeArgumentProvider
 import org.jetbrains.intellij.platform.gradle.artifacts.transform.ExtractorTransformer
 import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformDependenciesExtension
@@ -408,6 +409,7 @@ internal fun <T : Task> Project.preconfigureTask(task: T) {
          */
         if (this is SplitModeAware) {
             splitMode.convention(extension.splitMode)
+            targetProductPart.convention(extension.targetProductPart)
         }
 
         /**
@@ -434,11 +436,13 @@ internal fun <T : Task> Project.preconfigureTask(task: T) {
                 )
             )
             jvmArgumentProviders.add(
-                SandboxArgumentProvider(
+                SandboxArgumentProviderSplitModeAware(
                     sandboxConfigDirectory,
                     sandboxPluginsDirectory,
                     sandboxSystemDirectory,
                     sandboxLogDirectory,
+                    splitMode,
+                    targetProductPart,
                 )
             )